Continue with the Home page admin setup.

+++ frontend/src/components/MediaLibrary.jsx
@@ -17,6 +17,7 @@ const MediaLibrary = ({ displayMedia, setDisplayMedia, setSelectedImg }) => {
   const uploadCallback = async (file) => {
     const formData = new FormData();
     formData.append('image', file);
+    console.log(formData);
     try {
       const res = await uploadImage(formData).unwrap();
       toast.success(res.message);
@@ -38,12 +39,15 @@ const MediaLibrary = ({ displayMedia, setDisplayMedia, setSelectedImg }) => {
   };
 
   const filesDeleteCallback = async (item) => {
-    try {
-      const res = await deleteImage(item[0]._id);
-      toast.success(res.message);
-      refetch();
-    } catch (err) {
-      toast.error(err?.data?.message || err.error);
+    if (window.confirm('Are you sure delete this image?')) {
+      try {
+        const res = await deleteImage(item[0]._id);
+        toast.success(res.data.message);
+        refetch();
+      } catch (err) {
+        console.log(err);
+        toast.error(err?.data?.message || err.error);
+      }
     }
   };
 
@@ -59,6 +63,7 @@ const MediaLibrary = ({ displayMedia, setDisplayMedia, setSelectedImg }) => {
         </Message>
       ) : (
         <ReactMediaLibrary
+          multiSelect
           isOpen={displayMedia}
           onClose={() => setDisplayMedia(false)}
           fileLibraryList={images.data}
